move code to optimization_handler.py
